% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynafluxr.R
\name{cli}
\alias{cli}
\title{Function to be called from shell command line}
\usage{
cli(args = commandArgs(trailingOnly = TRUE))
}
\description{
Function to be called from shell command line
}
\examples{
  ddir=system.file("data", package="dynafluxr")
  meas=file.path(ddir, "data_teusink.tsv")
  sto=file.path(ddir, "network_teusink.txt")
  res=cli(c("-m", meas, "-s", sto, "--skip", "10"))
  tp=res$mf$Time
  np=length(tp)
  npp=10L # number of plot points in each tp interval
  dtp=diff(tp)
  tpp=tp[1L]+c(0.,cumsum(rep(diff(tp)/npp, each=npp)))
  # plot metabolites
  matplot(tpp, res$msp(tpp), type="l")
  matpoints(tp, res$mf[,-1], pch=".")
  legend("topright", legend=colnames(bsppar(res$msp)$qw), lty=1:5, col=1:6)
  # plot fluxes
  dev.new()
  matplot(tpp, res$fsp(tpp), type="l")
  ref=t(read.table(file.path(ddir, "glyco_teusink.flux.tsv"), sep="\t", header=TRUE, row.names=1, check.names=FALSE))
  tf=as.numeric(rownames(ref)) # reference flux time points
  nm_flux=colnames(bsppar(res$fsp)$qw)
  itf=(tf >= min(tp) & tf <= max(tp))
  matpoints(tf[itf], ref[itf, nm_flux, drop=FALSE], pch=".")
  legend("topright", legend=nm_flux, lty=1:5, col=1:6)
  # plot residuals
  dev.new()
  matplot(tpp, res$rsp(tpp), type="l")
  legend("topright", legend=colnames(bsppar(res$rsp)$qw), lty=1:5, col=1:6)
}
