% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynafluxr.R
\name{fdyn}
\alias{fdyn}
\title{Retrieve flux dynamics from metabolic kinetics}
\usage{
fdyn(mf, sto, nsp = 4L, nki = 5L, lieq = NULL)
}
\value{

}
\description{
Retrieve flux dynamics from metabolic kinetics
}
