% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt2sto.R
\name{txt2sto}
\alias{txt2sto}
\title{Translate plain text file with reactions to stoichiometric matrix}
\usage{
txt2sto(fn)
}
\arguments{
\item{fn}{String, file name}
}
\value{
Matrix of size n_metab x n_reac
}
\description{
Translate plain text file with reactions to stoichiometric matrix
}
