% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynafluxr.R
\name{fdyn}
\alias{fdyn}
\title{Retrieve flux dynamics from metabolic kinetics}
\usage{
fdyn(mf, sto, nsp = 4L, nki = 5L, lieq = NULL, monotone = 0, ils = FALSE)
}
\arguments{
\item{mf}{Data-frame or matrix, metabolite kinetic measurements.
Columns must be named with metabolite names and 'Time'.}

\item{sto}{Stoichiometric matrix, \code{sto[i,j]} means reaction 'j' produces
metabolite 'i' with the flux 'sto[i,j]'. If \code{sto[i,j] < 0},
the metabolite 'i' is consumed. Columns must be named with flux names.
Rows must be names with metabolite names.}

\item{nsp}{Integer, polynomial order of B-spline to use for metabolites}

\item{nki}{Integer, number of internal knots for B-splines}

\item{lieq}{List, equality constraints on metabolites}

\item{monotone}{Numeric scalar or vector, 1=metabolites are
monotonically increasing;
-1=monotonically decreasing; 0=no constraint. If vector, each value
constraints (or not) a corresponding data column in mf ('Time'
column is excluded
from counting)}
}
\value{
List with following components:
\describe{
  \item{mf:}{ metabolite data frame used for fitting}
  \item{sto:}{ stoichiometric matrix used for fitting}
  \item{msp:}{ metabolite spline function}
  \item{fsp:}{ flux spline function}
  \item{dsp:}{ metabolite first derivative spline function}
  \item{rsp:}{ residual \code{dm/dt - sto\%*\%flux} spline function}
}
}
\description{
Retrieve flux dynamics from metabolic kinetics
}
\details{
Each item in \code{lieq} corresponds to a metabolite and is a
  2 column matrix (Time, Value). Each
  row of this matrix indicates what 'Value' must take corresponding
  metabolite at what 'Time'. Typically, it can be used to impose
  starting values at Time=0 for some metabolites.\cr
  All metabolite fits are constraint to have values >= 0.
}
