% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt2sto.R
\name{txt2sto}
\alias{txt2sto}
\title{parse txt with stoichiometric equations
return stoichiometric matrix}
\usage{
txt2sto(fn)
}
\description{
parse txt with stoichiometric equations
return stoichiometric matrix
}
